backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0
rank-bm25==0.2.2
numpy==1.26.4
python-dotenv==1.0.1
langchain-text-splitters==0.2.2
pypdf==4.3.1
pydantic==2.7.4
requests==2.32.3